etl-processor/requirements.txt
pandas==2.2.3
numpy==2.1.2
influxdb-client==1.47.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.3.3
loguru==0.7.2
